Rename undoable => undo-redo (for consistency)
