+++ src/actions/index.ts
@@ -1,8 +1,9 @@
 import {ShelfAction} from './shelf';
-import {UndoableAction} from './undoable';
+import {UndoableAction} from './undo-redo';
 
 export * from './shelf';
 export * from './redux-action';
+export * from './undo-redo';
 
 /**
  * Union type of all actions in our application.
