+++ src/reducers/index.ts
@@ -1,11 +1,10 @@
 import undoable from 'redux-undo';
 
-import {Action} from '../actions';
+import {Action, UNDO, REDO} from '../actions';
 import {StateBase} from '../models';
 
 import {datasetReducer} from './dataset';
 import {shelfReducer} from './shelf';
-import {UNDO, REDO} from '../actions/undoable';
 import {HISTORY_LIMIT} from '../constants';
 
 function reducer(state: Readonly<StateBase>, action: Action): StateBase {
